Adjustments for Windows build with MSVC2019 (VC 16.10.31419.357)
1. removed some duplicate symbol linker error
2. removed many compiler warnings (mainly size_t/int compatibility)
3. consistent definition of db::pcell_id_type
4. removed UTF-8 character codes from string constants
5. float constants for float arguments
6. timeout in tlHttp when no openssl lib is found (instead of stalling
   app)

+++ src/db/db/dbCompoundOperation.cc
@@ -981,9 +981,9 @@ CompoundRegionLogicalCaseSelectOperationNode::result_type () const
   ResultType result = Region;
   for (size_t i = 1; i < children (); i += 2) {
     if (i == 1) {
-      result = child (i)->result_type ();
+      result = child ((unsigned int) i)->result_type ();
     } else {
-      tl_assert (result == child (i)->result_type ());
+      tl_assert (result == child ((unsigned int) i)->result_type ());
     }
   }
   return result;
@@ -1136,9 +1136,9 @@ CompoundRegionJoinOperationNode::result_type () const
   ResultType result = Region;
   for (size_t i = 0; i < children (); ++i) {
     if (i == 0) {
-      result = child (i)->result_type ();
+      result = child ((unsigned int) i)->result_type ();
     } else {
-      tl_assert (result == child (i)->result_type ());
+      tl_assert (result == child ((unsigned int) i)->result_type ());
     }
   }
   return result;
